@@ -1,54 +1,100 @@
 {
-   "name": "svps",
-   "version": "1.2.1",
-   "main": "./bin/index.js",
-   "dependencies": {
-      "async": "^3.2.4",
-      "ssh2": "^1.11.0"
-   },
-   "devDependencies": {
-      "eslint": "^8.31.0",
-      "packages-update": "^1.2.1"
-   },
-   "bin": {
-      "svps": "bin/index.js"
-   },
-   "scripts": {
-      "patch": "npm run update; npm audit fix; npm version patch --no-git-tag-version;",
-      "minor": "npm run update; npm audit fix; npm version minor --no-git-tag-version;",
-      "major": "npm run update; npm audit fix; npm version major --no-git-tag-version;",
-      "update": "npx npu; npm i --ignore-scripts;"
-   },
-   "author": "https://github.com/wellwelwel",
-   "license": "MIT",
-   "description": "🚀 A simple CLI tool to automate the setup and pre-settings of your Ubuntu VPS",
-   "engines": {
-      "node": ">=14.5.0",
-      "npm": ">=7.0.2"
-   },
-   "repository": {
-      "type": "git",
-      "url": "git+https://github.com/wellwelwel/svps.git"
-   },
-   "bugs": {
-      "url": "https://github.com/wellwelwel/svps/issues"
-   },
-   "homepage": "https://github.com/wellwelwel/svps#readme",
-   "keywords": [
-      "apache",
-      "cli",
-      "cronjobs",
-      "crontab",
-      "ufw",
-      "vsftpd",
-      "mysql",
-      "php",
-      "nodejs",
-      "sh",
-      "ssh",
-      "ubuntu",
-      "virtualhost",
-      "vps",
-      "vps-setup"
-   ]
+  "name": "svps",
+  "version": "2.0.0",
+  "type": "module",
+  "main": "./lib/index.js",
+  "bin": {
+    "svps": "bin/index.js"
+  },
+  "dependencies": {
+    "@types/ssh2": "^1.11.13",
+    "ssh2": "^1.14.0"
+  },
+  "devDependencies": {
+    "@types/node": "^20.4.5",
+    "@typescript-eslint/eslint-plugin": "^6.2.0",
+    "@typescript-eslint/parser": "^6.2.0",
+    "eslint": "^8.46.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.28.0",
+    "packages-update": "^1.2.1",
+    "prettier": "^3.0.0",
+    "typescript": "^5.1.6"
+  },
+  "scripts": {
+    "build": "rm -rf ./bin ./lib ./tests && npx tsc",
+    "build:examples": "npx tsc -p examples/tsconfig.json",
+    "patch": "npm run update; npm audit fix; npm version patch --no-git-tag-version",
+    "minor": "npm run update; npm audit fix; npm version minor --no-git-tag-version",
+    "major": "npm run update; npm audit fix; npm version major --no-git-tag-version",
+    "update": "npx npu; npm i --ignore-scripts",
+    "eslint:checker": "npx eslint . --ext .js,.ts",
+    "eslint:fix": "npx eslint . --fix",
+    "prettier:checker": "npx prettier --check .",
+    "prettier:fix": "npx prettier --write .github/workflows/*.yml resources/local-module .",
+    "docker:reset": "docker rm --force test-svps >/dev/null 2>&1 || true",
+    "docker:run": "docker run -d --name test-svps --privileged -p ${PORT:-22}:${PORT:-22} -e PORT=${PORT:-22} --restart always wellwelwel/vps:latest",
+    "tests:docker": "npm run docker:reset; npm run docker:run",
+    "tests:create": "cd ./tests && echo '{ \"type\": \"module\" }' | cat > ./package.json",
+    "tests:connection": "cd ./tests && node ./basic-connection.js",
+    "tests:mount": "cd ./tests && node ./mount.js",
+    "tests:commands": "cd ./tests && node ./commands.js",
+    "tests:upload": "cd ./tests && node ./upload.js",
+    "tests:virtual-hosts": "cd ./tests && node ./virtual-hosts.js",
+    "tests:reset": "npm run docker:reset",
+    "pretests": "npm run build && npm run build:examples && npm run eslint:checker && npm run prettier:checker && npm run tests:docker",
+    "tests": "npm run tests:create && npm run tests:connection && npm run tests:mount && npm run tests:commands && npm run tests:upload && npm run tests:virtual-hosts",
+    "posttests": "npm run tests:reset"
+  },
+  "author": "https://github.com/wellwelwel",
+  "license": "MIT",
+  "description": "🚀 An easier tool to automate your Ubuntu Server setup and domain forwarding",
+  "engines": {
+    "node": ">=14.5.0",
+    "npm": ">=7.0.2"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/wellwelwel/svps.git"
+  },
+  "bugs": {
+    "url": "https://github.com/wellwelwel/svps/issues"
+  },
+  "homepage": "https://github.com/wellwelwel/svps#readme",
+  "keywords": [
+    "vps",
+    "kvm",
+    "setup",
+    "ssh",
+    "sftp",
+    "ssh2",
+    "ubuntu",
+    "server",
+    "virtual",
+    "host",
+    "hosts",
+    "hosting",
+    "proxy",
+    "forward",
+    "domain",
+    "domains",
+    "rdp",
+    "remote",
+    "desktop",
+    "protocol",
+    "devops",
+    "cronjob",
+    "crontab",
+    "apache",
+    "ufw",
+    "vsftpd",
+    "mysql",
+    "php",
+    "docker"
+  ],
+  "files": [
+    "bin",
+    "lib",
+    "resources"
+  ]
 }
